Cleans up codebase.
