+++ SkateConnect/SkateConnect/Managers/Network.swift
@@ -12,55 +12,6 @@ import NostrSDK
 import SwiftUI
 import SwiftData
 
-func shareVideo(_ videoUrl: URL) {
-    // Implementation for sharing the video
-    print("Sharing video URL: \(videoUrl)")
-
-    if let url = URL(string: videoUrl.absoluteString) {
-        let fileNameWithoutExtension = url.deletingPathExtension().lastPathComponent
-        print("File name without extension: \(fileNameWithoutExtension)")
-        
-        // Construct the custom URL
-        let customUrlString = "\(Constants.LANDING_PAGE_SKATEPARK)/video/\(fileNameWithoutExtension)"
-        
-        // Ensure it's a valid URL
-        if let customUrl = URL(string: customUrlString) {
-            print("Custom URL: \(customUrl)")
-            
-            // Open the custom URL
-            if UIApplication.shared.canOpenURL(customUrl) {
-                UIApplication.shared.open(customUrl, options: [:], completionHandler: nil)
-            } else {
-                print("Unable to open URL: \(customUrl)")
-            }
-        } else {
-            print("Invalid custom URL")
-        }
-    }
-}
-
-func shareChannel(_ channelId: String) {
-    // Implementation for sharing channel
-    print("Sharing channel with id: \(channelId)")
-    
-    // Construct the custom URL
-    let customUrlString = "\(Constants.LANDING_PAGE_SKATEPARK)/channel/\(channelId)"
-    
-    // Ensure it's a valid URL
-    if let customUrl = URL(string: customUrlString) {
-        print("Custom URL: \(customUrl)")
-        
-        // Open the custom URL
-        if UIApplication.shared.canOpenURL(customUrl) {
-            UIApplication.shared.open(customUrl, options: [:], completionHandler: nil)
-        } else {
-            print("Unable to open URL: \(customUrl)")
-        }
-    } else {
-        print("Invalid custom URL")
-    }
-}
-
 class Network: ObservableObject, RelayDelegate, EventCreating {
     @Published var relayPool: RelayPool?
     
@@ -142,6 +93,7 @@ class Network: ObservableObject, RelayDelegate, EventCreating {
             return
         }
     }
+    
     func relay(_ relay: Relay, didReceive event: RelayEvent) {}
     func relay(_ relay: Relay, didReceive response: RelayResponse) {
         DispatchQueue.main.async {
